added customer managment and role restrictions

+++ app/Traits/CrudOperationsTrait.php
@@ -3,17 +3,27 @@
 namespace App\Traits;
 
 use App\Constants\HttpResponse;
+use Illuminate\Support\Facades\Gate;
 
 trait CrudOperationsTrait
 {
     public function create($data): array
     {
-        $response =  $this->getModel()->create($data);
-        $sucesss['data'] = $response;
-        $sucesss['message'] = HttpResponse::DATA_CREATED_MESSAGE;
-        $sucesss['status_code'] = HttpResponse::CREATED_CODE;
-
-        return $sucesss;
+        if(Gate::allows('isOwner')){
+            $response =  $this->getModel()->create($data);
+            $sucesss['data'] = $response;
+            $sucesss['message'] = HttpResponse::DATA_CREATED_MESSAGE;
+            $sucesss['status_code'] = HttpResponse::CREATED_CODE;
+    
+            return $sucesss;
+        }
+        else{
+            $error['error'] = true;
+            $error['data'] = [];
+            $error['message'] = HttpResponse::ACCESS_DENIED_MESSAGE;
+            $error['status_code'] = HttpResponse::ACCESS_DENIED_CODE;
+            return $error;
+        }
     }
 
     public function show($id)
@@ -46,12 +56,43 @@ trait CrudOperationsTrait
 
     public function delete($id)
     {
+        if(!Gate::allows('isCashier')){
         $model = $this->getModel()->findOrFail($id);
         $model->delete();
         $sucesss['data'] = [];
         $sucesss['message'] = HttpResponse::DATA_REMOVED_MESSAGE;
+        $sucesss['status_code'] = HttpResponse::NO_CONTENT_CODE;
 
         return $sucesss;
+        }
+        else{
+            $error['error'] = true;
+            $error['data'] = [];
+            $error['message'] = HttpResponse::ACCESS_DENIED_MESSAGE;
+            $error['status_code'] = HttpResponse::ACCESS_DENIED_CODE;
+            return $error;
+        }
+    }
+
+    public function permanentlyDelete($id)
+    {
+        if(Gate::allows('isOwner')){
+        $model = $this->getModel()->withTrashed()->findOrFail($id);
+        $model->forcedelete();
+        $sucesss['data'] = [];
+        $sucesss['message'] = HttpResponse::DATA_REMOVED_MESSAGE;
+        $sucesss['status_code'] = HttpResponse::NO_CONTENT_CODE;
+
+        return $sucesss;
+        }
+        else{
+            $error['error'] = true;
+            $error['data'] = [];
+            $error['message'] = HttpResponse::ACCESS_DENIED_MESSAGE;
+            $error['status_code'] = HttpResponse::ACCESS_DENIED_CODE;
+            return $error;
+        }
+        
     }
 
     abstract protected function getModel();
